refactoring the structure
