+++ ecommerce/ecommerce/urls.py
@@ -37,14 +37,16 @@ schema_view = get_schema_view(
 )
 
 urlpatterns = [
-    path('swagger/', schema_view.with_ui(cache_timeout=0), name='schema-json'),
-    path('', include('users.urls')),
     path('api-auth/', include('rest_framework.urls')),
+    path('swagger/', schema_view.with_ui(cache_timeout=0), name='schema-json'),
     path('admin/', admin.site.urls),
+    path('', include('users.urls.template_urls')),
+    path('user/v1/api/', include('users.urls.api_url')),
     path('blog/', include('post.urls')),
-    path('shop/', include('shops.urls')),
+    path('shop/', include('shops.urls.template_urls')),
+    path('shop/v1/api/' ,  include('shops.urls.api_urls')),
     path('product/', include('products.urls')),
-    path('order/', include('orders.urls')),
+    path('order/', include('orders.urls.template_url')),
+    path('order/v1/api/', include('orders.urls.api_url')),
 
-    
 ]+static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
