+++ ecommerce/shops/urls/template_urls.py
@@ -1,7 +1,6 @@
 from os import name
 from django.urls import path
-from . import views
-
+from shops.views import template_views as views
 
 app_name = 'shops'
 urlpatterns = [
@@ -10,7 +9,4 @@ urlpatterns = [
     path('dashboard/delete_shop/<slug:slug>/' ,views.RenderDeleteShop.as_view()  ,name="delete_shop_admin" ),
     path('dashboard/create_shop/' ,views.CreateShop.as_view()  ,name="create_shop" ),
     path('dashboard/edit_shop/<slug:slug>/' ,views.EditShop.as_view()  ,name="edit_shop" ),
-    path('listofshops/api/' ,views.ListOfShopsForCustomer.as_view() , name="list_of_shops"),
-    path('list/type/shops/' ,views.TypeOfShops.as_view() , name='type_of_shops'),
-    path('shop/<slug:shop_slug>/product/',views.ProductOfShop.as_view() , name="product_of_shop"),
 ]
