+++ ecommerce/users/models.py
@@ -20,12 +20,13 @@ class CustomUser(AbstractUser):
     REQUIRED_FIELDS = ['email']
     user_type =models.CharField(choices=USER_TYPE, default=CUSTOMER, max_length=2,) 
     # email = models.EmailField(unique=True)
-    phone = models.CharField(unique=True,max_length=20 , null=True ,blank=True)
+    phone = models.CharField(unique=True,max_length=20 )
     address =  models.CharField(max_length=100 , null=True ,blank=True)
-    created_at = models.DateTimeField(default=datetime.now, blank=True)
+    created_at = models.DateTimeField(auto_now_add=True, blank=True)
     city=models.CharField(max_length=25 , null=True ,blank=True)
     zip=models.CharField(max_length=100 , null=True ,blank=True)
     image = models.ImageField(upload_to='images', blank=True, null=True)
+    is_register=models.BooleanField(default=False)
 
     objects = CustomUserManager()
